+++ render/parameter.c
@@ -29,7 +29,7 @@ render_parameterbuffer_allocate(render_backend_t* backend, render_usage_t usage,
 	                                                   sizeof(render_parameterbuffer_t) + paramsize, 0,
 	                                                   MEMORY_PERSISTENT | MEMORY_ZERO_INITIALIZED);
 	buffer->backend    = backend;
-	buffer->usage      = usage;
+	buffer->usage      = (uint8_t)usage;
 	buffer->buffertype = RENDERBUFFER_PARAMETER;
 	buffer->policy     = RENDERBUFFER_UPLOAD_ONDISPATCH;
 	buffer->size       = data_size;
