+++ render/state.c
@@ -27,7 +27,7 @@ render_statebuffer_allocate(render_backend_t* backend, render_usage_t usage,
 	                                               sizeof(render_statebuffer_t), 0,
 	                                               MEMORY_PERSISTENT | MEMORY_ZERO_INITIALIZED);
 	buffer->backend    = backend;
-	buffer->usage      = usage;
+	buffer->usage      = (uint8_t)usage;
 	buffer->buffertype = RENDERBUFFER_STATE;
 	buffer->policy     = RENDERBUFFER_UPLOAD_ONDISPATCH;
 	buffer->size       = sizeof(render_state_t);
